+++ BASE.HPP
@@ -3,37 +3,105 @@
 
 const int K = 14;
 typedef char str[80];
-struct date
-{
-	int day;
-	int month;
-	int year;
+typedef char phn[11];
+
+enum sex {MALE, FEMALE, TRANS};
+enum date_type {DAY, MONTH, YEAR};
+enum time_type {HOUR, MINUTE, SECOND};
+enum body_parts {BRAIN, HEART, SKIN, LUNG, BONE, GEN}; // GEN for general problems
+enum adr_parts {HOUSE_NO, STREET, CITY, DISTRICT, STATE};
+enum times_of {START, END};
+
+struct time{
+	unsigned int hour;
+	unsigned int minute;
+	unsigned int second;
+
+	time(){
+		hour = 0;
+		minute = 0;
+		second = 0;
+	}
+	time(unsigned h, unsigned m, unsigned s){
+		if(h < 24 && m < 60 && s < 60){
+			hour = h;
+			minute = m;
+			second = s;
+		}
+		else{
+			hour = 0;
+			minute = 0;
+			second = 0;
+		}
+	}
 };
 
-struct time
-{
-	int hour;
-	int min;
-	int sec;
+struct date{
+	unsigned int day;
+	unsigned int month;
+	unsigned int year;
+
+	date(){
+		day = 0;
+		month = 0;
+		year = 0;
+	};
+
+	date(unsigned d, unsigned m, unsigned y){
+		if( d<=31 && m <=12){
+			day = d;
+			month = m;
+			year = y;
+		}
+		else{
+			day = 0;
+			month = 0;
+			year = 0;
+		}
+	}
+
 };
 
-struct address
-{
-	str state;
-	str district;
-	str city;
-	str street;
+struct adr{
 	str house_no;
+	str street;
+	str city;
+	str district;
+	str state;
 };
 
-struct insurance
-{
+struct disease{
+	str name;
+	int type;  		// refers to body part affected (LUNG, HEART, etc)
+};
+
+struct insurance{
 	str provider;
-	unsigned long int amount;
-	unsigned long int id_no;
-	date expiry_date;
+	unsigned long amount;
+	date expiry;
 };
 
+class person{
+	public:
+		person(str, int, date, adr, phn);	// Explicit constructor
+
+		// 'Get's
+		char* getname();
+		int getage();
+		int getsex();
+		int getdob(int);
+		char* getadr(int);
+		char* getphn();
+
+	protected:
+		str name;
+		unsigned age;
+		unsigned sex;
+		date dob;
+		adr address;
+		phn phone;
+
+};
 
 
 #endif
